Ajuste formato horas clientes exbidos na paginal inicial - grafico

+++ WebAppSystems/Models/ChartData.cs
@@ -3,6 +3,6 @@
     public class ChartData
     {
         public List<string> ClientNames { get; set; }
-        public List<int> ClientValues { get; set; }
+        public List<double> ClientValues { get; set; } // Alterado para List<double>
     }
 }
